// frontend/src/backend.d.ts
export interface RoomType {
    features: Array<string>;
    imageUrls: Array<string>;
    pricePerNight: bigint;
    name: string;
    offerPrice: bigint;
}